+++ alias-0x02.c
