+++ command-0x00.c
@@ -1,7 +1,4 @@
-#include "simple_shell.h"
-#include "stdsh.h"
 #include "command.h"
-
