+++ command-0x02.c
@@ -1,15 +1,11 @@
 #include <libgen.h>
 #include <limits.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 #include <sys/wait.h>
 
 #include "alias.h"
 #include "command.h"
+#include "shutils.h"
 #include "simple_shell.h"
-#include "stdsh.h"
-#include "utils.h"
