+++ shutils-0x00.c
@@ -1,8 +1,4 @@
-#include "stdsh.h"
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <string.h>
+#include "shutils.h"
 
 /**
  * _strlen - Calculate the length of a string.
