+++ shutils-0x03.c
@@ -1,8 +1,6 @@
-#include "simple_shell.h"
-#include "stdsh.h"
-#include "string.h"
+#include "shutils.h"
+#include "constants.h"
 
-#include <stdbool.h>
 
 /**
  * count_tokens - Count the number of tokens in a string
@@ -82,7 +80,7 @@ void sanitize_input(char *input)
 	char *src;
 
 	char *dest = input;
-	bool prev_space = false;
+	int prev_space = 0;
 
 	for (src = input; *src != '\0'; ++src)
 	{
@@ -91,13 +89,13 @@ void sanitize_input(char *input)
 			if (!prev_space)
 			{
 				*dest++ = ' ';
-				prev_space = true;
+				prev_space = 1;
 			}
 		}
 		else
 		{
 			*dest++ = *src;
-			prev_space = false;
+			prev_space = 0;
 		}
 	}
