+++ tests/test_alias.c
@@ -4,7 +4,7 @@
 #include <execinfo.h>
 
 #include "simple_shell.h"
-#include "stdsh.h"
+#include "shutils.h"
 
 shell_t *shell;
