--- utils.h
@@ -1,6 +1,0 @@
-#ifndef UTILS_H
-#define UTILS_H
-
-int is_absolute(char *path);
-
-#endif
